requirements.txt
python-dotenv
python-youtube
feedparser
numpy
requests
Pillow
reportlab
praw
beautifulsoup4
lxml
google-genai